Create Lab PCA_2_2_22_1_A
Estimated time

10 minutes
Level of difficulty

Easy
Objectives

Familiarize the student with:

    using the break statement in loops;
    reflecting real-life situations in computer code.

Scenario

The break statement is used to exit/terminate a loop.

Using the while loop, design a program that continuously asks the user to enter a secret word (e.g.,"You're stuck in an infinite loop! Enter a secret word to leave the loop:") unless the user enters "chupacabra" as the secret exit word, in which case the message "You've successfully left the loop" should be printed to the screen, and the loop should terminate.

Don't print any of the words entered by the user. Use the concept of conditional execution and the break statement.